requirements.txt
django
djangorestframework
instaloader
pytube